requirements.txt
pylab-sdk
matplotlib
pandas
numpy
jupyterlab
ipykernel